Staging progress for lunch
Thus far its been a matter of experimentation and understanding how to wishfully program the interactions in a way I might beable to use later. Currently working on a problematic default to None so can check for None in basecase, things must be different in the way its handled in python or I have a bug.